Moved weekday navbar into components dir
